Fix issue with session_id
