+++ projects/packages/woocommerce-analytics/src/class-woocommerce-analytics.php
@@ -42,12 +42,6 @@ class Woocommerce_Analytics {
 		// Initialize general store tracking actions.
 		add_action( 'init', array( new Universal(), 'init_hooks' ) );
 		add_action( 'init', array( new My_Account(), 'init_hooks' ) );
-		if (
-			class_exists( '\Automattic\WooCommerce\Blocks\Package' )
-			&& version_compare( \Automattic\WooCommerce\Blocks\Package::get_version(), '11.6.2', '>=' )
-		) {
-			add_action( 'init', array( new Checkout_Flow(), 'init_hooks' ) );
-		}
 
 		/**
 		 * Fires after the WooCommerce Analytics package is initialized
